chore: uninstall npm on ipfs
License: MIT
